# Optimize memory usage of SDS cluster config generation