```
$ benchstat /tmp/{old,new}
name                                 old time/op        new time/op        delta
ClusterGeneration/gateways-8               3.59ms ± 8%        3.79ms ±17%     ~     (p=0.548 n=5+5)
ClusterGeneration/gateways-shared-8        3.77ms ± 3%        3.83ms ±15%     ~     (p=0.548 n=5+5)
ClusterGeneration/empty-8                  1.69ms ±10%        1.80ms ±11%     ~     (p=0.421 n=5+5)
ClusterGeneration/tls-8                    5.91ms ±11%        5.18ms ± 2%  -12.34%  (p=0.016 n=5+4)
ClusterGeneration/telemetry-8              1.59ms ± 2%        1.62ms ± 4%     ~     (p=0.421 n=5+5)
ClusterGeneration/virtualservice-8          420µs ± 5%         391µs ± 2%   -6.91%  (p=0.008 n=5+5)

name                                 old alloc/op       new alloc/op       delta
ClusterGeneration/gateways-8               1.99MB ± 0%        1.99MB ± 0%     ~     (p=0.421 n=5+5)
ClusterGeneration/gateways-shared-8        1.99MB ± 0%        1.99MB ± 0%     ~     (p=0.222 n=5+5)
ClusterGeneration/empty-8                   819kB ± 0%         819kB ± 0%     ~     (p=1.000 n=5+5)
ClusterGeneration/tls-8                    2.54MB ± 0%        2.18MB ± 0%  -14.30%  (p=0.008 n=5+5)
ClusterGeneration/telemetry-8               820kB ± 0%         820kB ± 0%     ~     (p=0.556 n=4+5)
ClusterGeneration/virtualservice-8          202kB ± 0%         202kB ± 0%     ~     (p=0.452 n=5+5)

name                                 old allocs/op      new allocs/op      delta
ClusterGeneration/gateways-8                27.1k ± 0%         27.1k ± 0%     ~     (all equal)
ClusterGeneration/gateways-shared-8         27.1k ± 0%         27.1k ± 0%     ~     (all equal)
ClusterGeneration/empty-8                   11.3k ± 0%         11.3k ± 0%     ~     (all equal)
ClusterGeneration/tls-8                     34.5k ± 0%         28.1k ± 0%  -18.54%  (p=0.008 n=5+5)
ClusterGeneration/telemetry-8               11.3k ± 0%         11.3k ± 0%     ~     (p=0.444 n=5+5)
ClusterGeneration/virtualservice-8          2.81k ± 0%         2.81k ± 0%     ~     (all equal)
```