@howardjohn 2020-07-14
In a real cluster 10k pods, this was 5.39% of allocs