@istio-testing 2020-07-14
@howardjohn: The following tests **failed**, say `/retest` to rerun all failed tests:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
unit-tests_istio | acddf2f8d2b73799af74b9462929d28a36306d1b | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/25511/unit-tests_istio/16592) | `/test unit-tests_istio`
integ-operator-controller-tests_istio | 246d5df3cd86e2276d9e046af5f9da58054d8fc1 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/25511/integ-operator-controller-tests_istio/4412) | `/test integ-operator-controller-tests_istio`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->